chore: extend the 150-line budget to the platform shells
`scripts/check-loc.sh` only looked at `crates/*/src`. The platform shells are
excluded from the cargo workspace — each links a heavy OS-only dependency — so
nothing looked at them at all, which is how they drifted to a 550-line
`shell.rs` while every crate stayed inside the budget. Widen the find to any
`src` dir under `platforms/` (not `platforms/*/src`, which would miss
`platforms/linux/settings-gtk/src`), and split what that turns red:

- `background/hook.rs` (339) → one module per OS callback: `keyboard`, `mouse`,
  `foreground`, with the message pump in `mod.rs`. The two identical
  `send_plan_primed`/`send_plan` branches collapse into one `send` helper.
- `shared/commands.rs` (207) → `settings` / `system` / `config` / `updates`,
  along the section banners that were already there.
- `shared/update.rs` (214) → `feed` (network) and `install` (verify + swap),
  with the manifest types and version comparison staying in `mod.rs`.
- `shared/shell.rs` (198) → what the hook calls stays in `mod.rs`; what the
  Settings window calls moves to `shell/settings.rs`.
- `ui/lifecycle/child.rs` (152) → the Control Center's toggle/reap rules move to
  `lifecycle/flyout.rs`; it is the one child with a debounce and an exit-code
  protocol, and the generic spawn/kill path reads better without it.
- `ui/recorder.rs` (152) → the VK lookup tables move to `recorder/keys.rs`.
- `crates/funput-core/src/unicode/shapes.rs` (155) → the shaped-vowel table moves
  to `shapes/table.rs`. This one was **already failing** before this branch, so
  the gate had to be fixed either way.

`shell` and `commands` keep byte-identical public APIs (47 and 28 items), so no
caller changes. Verified as before against `x86_64-pc-windows-gnu`; `commands`
and `ui/` are the parts that need slint and stay uncovered.
